フロントエンド実装

+++ pages/api/item/delete/[id].ts
@@ -1,11 +1,14 @@
+import type { NextApiResponse } from "next";
 import connectDB from "../../../../utils/database";
 import { ItemModel } from "../../../../utils/schemaModels";
 import auth from "../../../../utils/auth";
+import { ExtendedNextApiRequestItem, SavedItemDataType, ResMessageType } from "../../../../utils/types";
 
-const deleteItem = async (req, res) => {
+const deleteItem = async (req: ExtendedNextApiRequestItem, res: NextApiResponse<ResMessageType>) => {
   try {
     await connectDB();
-    const singleItem = await ItemModel.findById(req.query.id);
+    const singleItem: SavedItemDataType | null = await ItemModel.findById(req.query.id);
+    if (!singleItem) return res.status(400).json({ message: "アイテムが存在していないため削除失敗" });
     if (singleItem.email === req.body.email) {
       await ItemModel.deleteOne({ _id: req.query.id });
       return res.status(200).json({ message: "アイテム削除成功" });
